First parsing of the log file in place. Some objects do not match because the log file is made with OpenPilot firmware, and a different version. These XMLs are taken from the Taulabs version.
Need to retrieve the XML definition files from the exac same version of the firmware in teh board right now to see no errors. 

Otherwise, for those objects recognized, it's working apparentyl fine. This is expected, since the hash function performs a bit of a check on our software. 

Still not sure, however, that the log file contains all the communication with the embedded firmware (may be well recording just a few packets marked with the logging attribute)